Refactor infrastructure legend into its own component
